rewrite to debian

+++ apps/zsh.sh
@@ -1,48 +1,41 @@
 #!/bin/bash
 
+set -e
+
 source "$SHELLSMITH_UTILS/safe_symlink.sh"
 
-install_and_setup_zsh() {
-  sudo apt install zsh -y
-  chsh -s /usr/bin/zsh
+install_zsh() {
+  sudo apt-get update
+  sudo apt-get install -y zsh git bat
+  sudo chsh -s /usr/bin/zsh $(logname)
 }
 
-install_and_setup_ohmyzsh() {
-  if [ -d "$HOME/.oh-my-zsh/" ]; then
-    rm -rf "$HOME/.oh-my-zsh/"
-  fi
-
-  sh -c "$(curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)" "" --unattended
-
+install_ohmyzsh() {
+  [ -d "$HOME/.oh-my-zsh" ] && rm -rf "$HOME/.oh-my-zsh"
+  git clone --depth=1 https://github.com/ohmyzsh/ohmyzsh.git "$HOME/.oh-my-zsh"
   safe_symlink "$SHELLSMITH_DOTFILES/zshrc" "$HOME/.zshrc"
 }
 
 install_plugins() {
-  sudo apt install bat
+  sudo apt-get install -y bat
+  local plugins_dir="${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/plugins"
 
-  local plugins_dir=${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins
+  rm -rf "$plugins_dir"
+  mkdir -p "$plugins_dir"
 
-  rm -rf "$plugins_dir" && mkdir -p "$plugins_dir"
-
-  git clone https://github.com/zsh-users/zsh-autosuggestions \
-    "$plugins_dir/zsh-autosuggestions"
-  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git \
-    "$plugins_dir/zsh-syntax-highlighting"
-  git clone https://github.com/fdellwing/zsh-bat.git \
-    "$plugins_dir/zsh-bat"
+  git clone --depth=1 https://github.com/zsh-users/zsh-autosuggestions "$plugins_dir/zsh-autosuggestions"
+  git clone --depth=1 https://github.com/zsh-users/zsh-syntax-highlighting.git "$plugins_dir/zsh-syntax-highlighting"
+  git clone --depth=1 https://github.com/fdellwing/zsh-bat.git "$plugins_dir/zsh-bat"
 }
 
-install_and_setup_p10k() {
-  local p10k_dir=${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k
+install_p10k() {
+  local p10k_dir="${ZSH_CUSTOM:-$HOME/.oh-my-zsh/custom}/themes/powerlevel10k"
 
-  if [ -d "$p10k_dir" ]; then
-    rm -rf "$p10k_dir"
-  fi
-
+  [ -d "$p10k_dir" ] && rm -rf "$p10k_dir"
   git clone --depth=1 https://github.com/romkatv/powerlevel10k.git "$p10k_dir"
 }
 
-install_and_setup_zsh
-install_and_setup_ohmyzsh
+install_zsh
+install_ohmyzsh
 install_plugins
-install_and_setup_p10k
+install_p10k
